# public/swagger_v4.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "https://api.bible.build/v2",
      "description": "Online Staging server for testing"
    },
    {
      "url": "https://api.dbp.dev/v2",
      "description": "Internal Staging server for testing"
    }
  ],
  "info": {
    "title": "Digital Bible Platform [Koinos]",
    "description": "This is a simple API",
    "version": "4.0.0",
    "contact": {
      "name": "API Support",
      "url": "https://dbs.org/contact",
      "email": "jon@dbs.org"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/bibles/filesets/{id}/download": {
      "get": {
        "operationId": "v4_bible_filesets.download",
        "summary": "Download either a full zipped audio Bible or a selection",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          },
          {
            "name": "bucket_id",
            "in": "query",
            "required": true,
            "description": "This param states which bucket the resource should be downloaded from.",
            "schema": {
              "type": "string",
              "maxLength": 12,
              "example": "dbs-web"
            }
          }
        ],
        "tags": [
          "Bible"
        ],
        "responses": {
          "200": {
            "description": "An array of Bibles",
            "content": {
              "application/zip": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "example": "file.zip"
                }
              }
            }
          }
        }
      }
    },
    "/bibles/filesets/{id}/podcast": {
      "get": {
        "operationId": "v4_bible.podcast",
        "summary": "An rss ready format for the audio filesets",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Bible"
        ],
        "responses": {
          "200": {
            "description": "An podcastable ready format for the audio filesets",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/languages": {
      "get": {
        "operationId": "v4_languages.all",
        "summary": "Get List of Languages",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Wiki"
        ],
        "responses": {
          "200": {
            "description": "An array of Languages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Languages"
                }
              }
            }
          }
        }
      }
    },
    "/languages/{id}": {
      "get": {
        "operationId": "v4_languages.one",
        "summary": "Get a specific Language",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Wiki"
        ],
        "responses": {
          "200": {
            "description": "An array of Languages",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Languages"
                }
              }
            }
          }
        }
      }
    },
...
    "/resources/{resource_id}": {
      "get": {
        "operationId": "v4_resources.show",
        "summary": "Get a single resource by it's ID",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Organizations"
        ],
        "responses": {
          "200": {
            "description": "A success resource index query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "v4_resources.update",
        "summary": "Update a single resource by it's ID.",
        "description": "This route will update a single resource by it's ID but in order to do so you must either belong to the organization that owns the resource or have `archivist` level permissions. Some resources are updated via API. Changes made to this type of resource will be overwritten on the next update. These resources are indentified by a `source_id`",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Organizations"
        ],
        "responses": {
          "200": {
            "description": "A successful resource update query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "v4_resources.store",
        "summary": "Create a single resource",
        "description": "This route will create a single resource but order to do so you must either belong to an organization or have `archivist` level permissions. Please do not create a resource for which your organization already has a connected API as that will introduce duplicates into the system. To see a list of connected APIs, please see the `connections` route.",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Organizations"
        ],
        "responses": {
          "200": {
            "description": "A successful resource creation query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "v4_resources.destroy",
        "summary": "Delete a Single Resource",
        "description": "This route will delete a single resource but order to do so you must either belong to the owning organization or have `archivist` level permissions.",
        "parameters": [
          {
            "$ref": "#/components/parameters/v"
          },
          {
            "$ref": "#/components/parameters/key"
          },
          {
            "$ref": "#/components/parameters/version"
          },
          {
            "$ref": "#/components/parameters/reply"
          },
          {
            "$ref": "#/components/parameters/pretty"
          }
        ],
        "tags": [
          "Organizations"
        ],
        "responses": {
          "200": {
            "description": "A successful deletion query",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resource"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Languages": {
        "required": [
          "name"
        ],
        "properties": {
          "glotto_code": {
            "type": "string",
            "example": "aari1239",
            "minLength": 8,
            "maxLength": 8
          },
          "iso_code": {
            "type": "string",
            "example": "aiw",
            "minLength": 3,
            "maxLength": 3
          },
          "name": {
            "type": "string",
            "example": "Aari"
          },
          "bible_count": {
            "type": "integer",
            "example": 3
          }
        }
      },
...
    },
    "parameters": {
      "v": {
        "name": "v",
        "in": "query",
        "description": "The Version Number of the API",
        "required": true,
        "example": 4,
        "schema": {
          "type": "string"
        }
      },
      "key": {
        "name": "key",
        "in": "query",
        "description": "The API key that's been generated for you.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "pretty": {
        "name": "pretty",
        "in": "query",
        "description": "A sugar param, simply will format the call for human eyes",
        "required": false,
        "schema": {
          "type": "string"
        }
      },
      "version": {
        "name": "key",
        "in": "query",
        "description": "The API key that's been generated for you.",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "reply": {
        "name": "reply",
        "in": "query",
        "description": "Specifies the response type requested by the caller. Default is json.",
        "schema": {
          "type": "string",
          "enum": [
            "json",
            "html",
            "jsonp",
            "yaml",
            "xml"
          ]
        }
      },
...
    }
  }
}